requirements.txt
pygame>=2.5.0
Pillow>=10.0.0
svglib>=1.5.1
reportlab>=4.0.4 